backend/requirements.txt
# ==========================================
# REQUIREMENTS.TXT - Dependências do Projeto
# ==========================================
# Para instalar: pip install -r requirements.txt
# ==========================================

# --- Flask e extensões ---
Flask==3.1.2
Werkzeug==3.1.3

# --- Banco de Dados ---
psycopg2-binary==2.9.11

# --- Segurança e Hash de Senhas ---
bcrypt==4.1.2

# --- Variáveis de Ambiente ---
python-dotenv==1.0.0

# --- Dependências do Flask (já instaladas com Flask) ---
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3

# --- Utilitários ---
importlib_metadata==8.7.0
zipp==3.23.0


# ==========================================
# DEPENDÊNCIAS DE SEGURANÇA E PRODUÇÃO
# ==========================================

# --- JWT para Autenticação ---
PyJWT==2.9.0
Flask-JWT-Extended==4.6.0

# --- CORS para permitir requisições de outros domínios ---
Flask-CORS==4.0.0

# --- Validação de dados ---
marshmallow==3.21.0

# --- Rate Limiting ---
Flask-Limiter==3.5.0

# --- Migrations de banco de dados ---
# Flask-Migrate==4.0.5
# alembic==1.12.1

# --- Testing ---
pytest==7.4.3
pytest-flask==1.3.0

# --- Produção (WSGI Server) ---
gunicorn==21.2.0

# --- API Documentation ---
flask-swagger-ui==4.11.1

# --- Email Validation ---
email-validator==2.1.0


# ==========================================
# NOTAS DE VERSÃO
# ==========================================
# Python requerido: >= 3.8
# PostgreSQL requerido: >= 12.0
# 
# Para desenvolvimento:
#   pip install -r requirements.txt
# 
# Para produção, adicione:
#   gunicorn
#   PyJWT
#   Flask-CORS
# ==========================================